[rs] Minor Rust fixes
- **[Breaking change]** Use `u16` for `GotoFrame` and `GotoFrame2` body.
- **[Feature]** Implement `Clone`, `Ord` and `Hash` for all types.
- **[Feature]** Implement `TryFrom` to convert from a CFG action to a raw action.
- **[Feature]** Expose `ActionHeader` type.
- **[Fix]** Fix JSON serialization for floating-point push values

+++ rs/src/action.rs
@@ -4,20 +4,20 @@ use crate::{GetUrl2Method, PushValue};
 #[cfg(feature = "use-serde")]
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, PartialEq, Eq)]
+#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 #[cfg_attr(feature = "use-serde", derive(Serialize, Deserialize))]
 pub struct ConstantPool {
   pub pool: Vec<String>,
 }
 
-#[derive(Debug, PartialEq, Eq)]
+#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 #[cfg_attr(feature = "use-serde", derive(Serialize, Deserialize))]
 pub struct GetUrl {
   pub url: String,
   pub target: String,
 }
 
-#[derive(Debug, PartialEq, Eq)]
+#[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 #[cfg_attr(feature = "use-serde", derive(Serialize, Deserialize))]
 pub struct GetUrl2 {
   pub method: GetUrl2Method,
@@ -25,32 +25,32 @@ pub struct GetUrl2 {
   pub load_variables: bool,
 }
 
-#[derive(Debug, PartialEq, Eq)]
+#[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 #[cfg_attr(feature = "use-serde", derive(Serialize, Deserialize))]
 pub struct GotoFrame {
-  pub frame: usize,
+  pub frame: u16,
 }
 
-#[derive(Debug, PartialEq, Eq)]
+#[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 #[cfg_attr(feature = "use-serde", derive(Serialize, Deserialize))]
 pub struct GotoFrame2 {
   pub play: bool,
-  pub scene_bias: usize,
+  pub scene_bias: u16,
 }
 
-#[derive(Debug, PartialEq, Eq)]
+#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 #[cfg_attr(feature = "use-serde", derive(Serialize, Deserialize))]
 pub struct GoToLabel {
   pub label: String,
 }
 
-#[derive(Debug, PartialEq, Eq)]
+#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 #[cfg_attr(feature = "use-serde", derive(Serialize, Deserialize))]
 pub struct Push {
   pub values: Vec<PushValue>,
 }
 
-#[derive(Debug, PartialEq, Eq)]
+#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 #[cfg_attr(feature = "use-serde", derive(Serialize, Deserialize))]
 pub struct Raw {
   pub code: u8,
@@ -61,19 +61,19 @@ pub struct Raw {
   pub data: Vec<u8>,
 }
 
-#[derive(Debug, PartialEq, Eq)]
+#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 #[cfg_attr(feature = "use-serde", derive(Serialize, Deserialize))]
 pub struct SetTarget {
   pub target_name: String,
 }
 
-#[derive(Debug, PartialEq, Eq)]
+#[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 #[cfg_attr(feature = "use-serde", derive(Serialize, Deserialize))]
 pub struct StoreRegister {
   pub register: u8,
 }
 
-#[derive(Debug, PartialEq, Eq)]
+#[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 #[cfg_attr(feature = "use-serde", derive(Serialize, Deserialize))]
 pub struct StrictMode {
   pub mode: u8,
